requirements.txt
python-dotenv==1.0.1
requests==2.31.0
pyyaml==6.0.2
pytest==8.4.2